+++ sdk/eventgrid/azure-messaging-eventgrid-systemevents/src/main/java/com/azure/messaging/eventgrid/systemevents/models/AcsRouterJobCompletedEventData.java
@@ -30,6 +30,18 @@ public final class AcsRouterJobCompletedEventData extends AcsRouterJobEventData 
     private String workerId;
 
     /*
+     * Router Jobs events Tags
+     */
+    @Generated
+    private Map<String, String> tags;
+
+    /*
+     * Router Job events Labels
+     */
+    @Generated
+    private Map<String, String> labels;
+
+    /*
      * Router Job events Queue Id
      */
     @Generated
@@ -51,12 +63,10 @@ public final class AcsRouterJobCompletedEventData extends AcsRouterJobEventData 
      * Creates an instance of AcsRouterJobCompletedEventData class.
      *
      * @param jobId the jobId value to set.
-     * @param labels the labels value to set.
-     * @param tags the tags value to set.
      */
     @Generated
-    private AcsRouterJobCompletedEventData(String jobId, Map<String, String> labels, Map<String, String> tags) {
-        super(jobId, labels, tags);
+    private AcsRouterJobCompletedEventData(String jobId) {
+        super(jobId);
     }
 
     /**
@@ -77,6 +87,28 @@ public final class AcsRouterJobCompletedEventData extends AcsRouterJobEventData 
     @Generated
     public String getWorkerId() {
         return this.workerId;
+    }
+
+    /**
+     * Get the tags property: Router Jobs events Tags.
+     *
+     * @return the tags value.
+     */
+    @Generated
+    @Override
+    public Map<String, String> getTags() {
+        return this.tags;
+    }
+
+    /**
+     * Get the labels property: Router Job events Labels.
+     *
+     * @return the labels value.
+     */
+    @Generated
+    @Override
+    public Map<String, String> getLabels() {
+        return this.labels;
     }
 
     /**
@@ -120,8 +152,6 @@ public final class AcsRouterJobCompletedEventData extends AcsRouterJobEventData 
     public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
         jsonWriter.writeStartObject();
         jsonWriter.writeStringField("jobId", getJobId());
-        jsonWriter.writeMapField("labels", getLabels(), (writer, element) -> writer.writeString(element));
-        jsonWriter.writeMapField("tags", getTags(), (writer, element) -> writer.writeString(element));
         jsonWriter.writeStringField("channelReference", getChannelReference());
         jsonWriter.writeStringField("channelId", getChannelId());
         jsonWriter.writeStringField("queueId", getQueueId());
@@ -174,7 +204,9 @@ public final class AcsRouterJobCompletedEventData extends AcsRouterJobEventData 
                 }
             }
             AcsRouterJobCompletedEventData deserializedAcsRouterJobCompletedEventData
-                = new AcsRouterJobCompletedEventData(jobId, labels, tags);
+                = new AcsRouterJobCompletedEventData(jobId);
+            deserializedAcsRouterJobCompletedEventData.labels = labels;
+            deserializedAcsRouterJobCompletedEventData.tags = tags;
             deserializedAcsRouterJobCompletedEventData.channelReference = channelReference;
             deserializedAcsRouterJobCompletedEventData.channelId = channelId;
             deserializedAcsRouterJobCompletedEventData.queueId = queueId;